Moving the "service" to be a close collaborator
